requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
Faker==12.0.1
idna==3.3
importlib-metadata==4.11.2
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
mixer==7.2.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.7.0
